Update chord_parser.dart
My proposed changes to properly fix the overflow issues. I have combined both the original function and my function.

The code has turned out much more complex than I was hoping as I kept running into little edge cases. Regardless, it seems to work better (ie no overflows errors) for me. I'm still not 100% sure that this is the best way to go forward with this but I need to commit it as it keeps playing on my mind! It can always be fixed later I guess.

I'm no flutter expert so I'm not sure on the correct way of naming variables (with the underscore etc) or writing comments.

I have also removed the need for the widgetPadding parameter by calculating the width directly using a LayoutBuilder in the lyrics_renderer file.